Changed macros SEC, MSEC, USEC (#801)

+++ Emulator/Base/Aliases.h
@@ -41,9 +41,9 @@ typedef u32 Block;
 //
 
 // Converts a certain unit to master cycles
-#define USEC(delay)           (Cycle(delay) * 28)
-#define MSEC(delay)           (Cycle(delay) * 28000)
-#define SEC(delay)            (Cycle(delay) * 28000000)
+#define USEC(delay)           (Cycle((i64)delay * 28))
+#define MSEC(delay)           (Cycle((i64)delay * 28000))
+#define SEC(delay)            (Cycle((double)delay * 28000000))
 
 #define CPU_CYCLES(cycles)    (Cycle(cycles) << 2)
 #define CIA_CYCLES(cycles)    (Cycle(cycles) * 40)
